requirements.txt
numpy==1.26.4
pandas==2.2.2
transformers==4.40.2
tensorflow==2.16.1
pymongo==4.7.2
langchain==0.1.20
wikipedia==1.4.0
streamlit==1.34.0
langchain_openai==0.1.6
python-dotenv==1.0.1